+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@
 * #64 Consul-based job locks
 * #65 Changed cron type from String to value object.
 * #66 Switch TriggerDescriptor properties to value objects
+* #72 Double --job option
 * #76 Scheduled job exceptions are not reported 
 
 ## 0.25